# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: VEEX Platform API
  description: |
    Enterprise-grade Industrial IoT Platform API for device management, OTA updates, and telemetry.
    
    ## Features
    - Device registry and fleet management
    - OTA campaign management with controlled rollouts
    - Real-time telemetry and analytics
    - Prometheus-compatible metrics
    
  version: 1.0.0
  contact:
    name: VEEX Platform
    url: https://github.com/veex-platform
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://registry.veex.dev
    description: Production server
...
paths:
...
  /api/v1/fleets:
    get:
      tags: [Fleets]
      summary: List fleets
      responses:
        '200':
          description: List of fleets
          content:
            application/json:
              schema:
                type: object
                properties:
                  fleets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Fleet'
                  total:
                    type: integer

    post:
      tags: [Fleets]
      summary: Create fleet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name]
              properties:
                id:
                  type: string
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Fleet created
        '409':
          description: Fleet ID already exists
  /api/v1/fleets/{id}:
    get:
      tags: [Fleets]
      summary: Get fleet details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fleet found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fleet'

    put:
      tags: [Fleets]
      summary: Update fleet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Fleet updated

    delete:
      tags: [Fleets]
      summary: Delete fleet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fleet deleted
  /api/v1/fleets/{id}/devices:
    post:
      tags: [Fleets]
      summary: Add device to fleet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id]
              properties:
                device_id:
                  type: string
      responses:
        '200':
          description: Device added to fleet

  # OTA Campaigns
  /api/v1/ota/campaigns:
    get:
      tags: [OTA]
      summary: List OTA campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, paused, completed, failed]
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaigns:
                    type: array
                    items:
                      $ref: '#/components/schemas/OTACampaign'

    post:
      tags: [OTA]
      summary: Create OTA campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name, artifact_id]
              properties:
                id:
                  type: string
                name:
                  type: string
                artifact_id:
                  type: string
                target_fleet_id:
                  type: string
      responses:
        '201':
          description: Campaign created
  /api/v1/ota/campaigns/{id}:
    get:
      tags: [OTA]
      summary: Get campaign details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Campaign found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OTACampaign'
  /api/v1/ota/campaigns/{id}/start:
    post:
      tags: [OTA]
      summary: Start campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Campaign started
  /api/v1/ota/campaigns/{id}/pause:
    put:
      tags: [OTA]
      summary: Pause campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Campaign paused
  /api/v1/ota/campaigns/{id}/resume:
    put:
      tags: [OTA]
      summary: Resume campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Campaign resumed

  # Analytics
...
components:
  schemas:
...
    Fleet:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        device_count:
          type: integer
    OTACampaign:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        artifact_id:
          type: string
        target_fleet_id:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed, failed]
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        success_count:
          type: integer
        failure_count:
          type: integer